+++ src/main/java/de/samply/db/repository/QueryExecutionRepository.java
@@ -1,0 +1,10 @@
+package de.samply.db.repository;
+
+import de.samply.db.model.QueryExecution;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QueryExecutionRepository extends JpaRepository<QueryExecution, Long> {
+
+}