+++ src/main/resources/db/migration/V001__init_schema_and_tables.sql
@@ -1,14 +1,38 @@
 CREATE SCHEMA IF NOT EXISTS samply;
 
+SET search_path TO samply;
+
 CREATE TABLE samply.query
 (
     id              SERIAL NOT NULL PRIMARY KEY,
-    query           text,
-    format          text,
-    label           text,
-    description     text,
-    contact_id      text,
-    expiration_date date,
-    created_at      timestamp,
-    archived_at     timestamp
+    query           TEXT,
+    format          TEXT,
+    label           TEXT,
+    description     TEXT,
+    contact_id      TEXT,
+    expiration_date DATE,
+    created_at      TIMESTAMP,
+    archived_at     TIMESTAMP
 );
+
+CREATE TABLE samply.query_execution
+(
+    id          SERIAL NOT NULL PRIMARY KEY,
+    query       TEXT,
+    query_id    SERIAL,
+    executed_at TIMESTAMP
+);
+
+CREATE INDEX idx_query_execution_query_id ON samply.query_execution (query_id);
+
+CREATE TABLE samply.query_execution_file
+(
+    id           SERIAL NOT NULL PRIMARY KEY,
+    query_execution_id SERIAL,
+    filePath     TEXT
+);
+
+CREATE INDEX idx_query_execution_file_query_execution_id ON samply.query_execution_file (query_execution_id);
+
+ALTER TABLE samply.query_execution ADD CONSTRAINT query_execution_query_id_fkey FOREIGN KEY ( query_id ) REFERENCES samply.query( id ) ON DELETE CASCADE;
+ALTER TABLE samply.query_execution_file ADD CONSTRAINT query_execution_file_query_execution_id_fkey FOREIGN KEY ( query_execution_id ) REFERENCES samply.query_execution( id ) ON DELETE CASCADE;